refactor(boards): unify P1/P2 HpsdrBoardKind into Zeus.Contracts (Phase 4)
Phase 4 of the radio-support plan. Promotes HpsdrBoardKind from per-
protocol Discovery namespaces (Zeus.Protocol1.Discovery and
Zeus.Protocol2.Discovery) to a single contract enum at
Zeus.Contracts.HpsdrBoardKind so server-side dispatch covers every
wire ID either protocol can emit, not just the P1 set.

Naming decisions for the unified enum (called out in
docs/designs/radio-support-plan.md §Phase 4):
  - 0x00 → Metis (Apache canonical name; P2's historical "Atlas"
    label collapsed in).
  - 0x02 → HermesII (matches Apache marketing copy and Thetis enum
    directly; P1's historical "Griffin" label collapsed in).

Wire-byte values are unchanged — old zeus-prefs.db rows hydrate
identically. Both ReplyParsers now emit the contract enum.

Files touched:
  - Zeus.Contracts/HpsdrBoardKind.cs — new canonical enum.
  - Zeus.Protocol1/Discovery/HpsdrBoardKind.cs — DELETED.
  - Zeus.Protocol2/Discovery/HpsdrBoardKind.cs — DELETED.
  - Both ReplyParsers + DiscoveredRadio + tests + every dispatch
    consumer (PaDefaults, RadioCalibrations, BoardCapabilitiesTable,
    RadioDriveProfile, RadioService, etc.) get `using Zeus.Contracts`.

Renames (test + dispatch update only — wire bytes stable):
  - HpsdrBoardKind.Griffin → HpsdrBoardKind.HermesII
  - HpsdrBoardKind.Atlas → HpsdrBoardKind.Metis

872 tests pass (same count as Phase 3). No behaviour change beyond P2
discoveries on wire 0x00 (Metis) now reaching server-side dispatch
(previously fell through to Unknown via the protocol enum split).

Refs #218.
