--- Zeus.Protocol1/Discovery/HpsdrBoardKind.cs
@@ -1,63 +1,0 @@
-// SPDX-License-Identifier: GPL-2.0-or-later
-//
-// Zeus — OpenHPSDR Protocol-1 / Protocol-2 client.
-// Copyright (C) 2025-2026 Brian Keating (EI6LF),
-//                         Douglas J. Cerrato (KB2UKA), and contributors.
-//
-// This program is free software: you can redistribute it and/or modify it
-// under the terms of the GNU General Public License as published by the
-// Free Software Foundation, either version 2 of the License, or (at your
-// option) any later version. See the LICENSE file at the root of this
-// repository for the full text, or https://www.gnu.org/licenses/.
-//
-// Zeus is an independent reimplementation in .NET — not a fork. Its
-// Protocol-1 / Protocol-2 framing, WDSP integration, meter pipelines, and
-// TX behaviour were informed by studying the Thetis project
-// (https://github.com/ramdor/Thetis), the authoritative reference
-// implementation in the OpenHPSDR ecosystem. Zeus gratefully acknowledges
-// the Thetis contributors whose work made this possible:
-//
-//   Richard Samphire (MW0LGE), Warren Pratt (NR0V),
-//   Laurence Barker (G8NJJ),   Rick Koch (N1GP),
-//   Bryan Rambo (W4WMT),       Chris Codella (W2PA),
-//   Doug Wigley (W5WC),        FlexRadio Systems,
-//   Richard Allen (W5SD),      Joe Torrey (WD5Y),
-//   Andrew Mansfield (M0YGG),  Reid Campbell (MI0BOT),
-//   Sigi Jetzlsperger (DH1KLM).
-//
-// Thetis itself continues the GPL-governed lineage of FlexRadio PowerSDR
-// and the OpenHPSDR (TAPR/OpenHPSDR) ecosystem; that lineage is preserved
-// here. See ATTRIBUTIONS.md at the repository root for the full provenance
-// statement and per-component attribution.
-//
-// Protocol-2 / PureSignal / Saturn-class behaviour was additionally informed
-// by pihpsdr (https://github.com/dl1ycf/pihpsdr), maintained by Christoph
-// Wüllen (DL1YCF); and by DeskHPSDR
-// (https://github.com/dl1bz/deskhpsdr), maintained by Heiko (DL1BZ).
-// Both are GPL-2.0-or-later.
-//
-// WDSP — loaded by Zeus via P/Invoke — is Copyright (C) Warren Pratt
-// (NR0V), distributed under GPL v2 or later.
-//
-// Zeus is distributed WITHOUT ANY WARRANTY; see the GNU General Public
-// License for details.
-
-namespace Zeus.Protocol1.Discovery;
-
-public enum HpsdrBoardKind : byte
-{
-    Metis = 0x00,
-    Hermes = 0x01,
-    Griffin = 0x02,
-    Angelia = 0x04,
-    Orion = 0x05,
-    HermesLite2 = 0x06,
-    OrionMkII = 0x0A,
-    // Apache Labs ANAN-G2E (N1GP firmware). Thetis HPSDRHW.HermesC10 — see
-    // enums.cs:400 and HardwareSpecific.Model setter at clsHardwareSpecific.cs:129
-    // in ramdor/Thetis. Single-RX, MKII-BPF, 33 mV ADC supply, no LR audio
-    // swap. Shares the OrionMkII / G2 PA forward-power calibration constants
-    // (bridge 0.12 / ref 5.0 / offset 32) per console.cs:25079-25088.
-    HermesC10 = 0x14,
-    Unknown = 0xFF,
-}
