+++ tests/Zeus.Protocol1.Tests/ControlFramePsEncoderTests.cs
@@ -13,6 +13,7 @@
 // See ATTRIBUTIONS.md at the repository root for the full provenance
 // statement and per-component attribution.
 
+using Zeus.Contracts;
 using Zeus.Protocol1.Discovery;
 
 namespace Zeus.Protocol1.Tests;
