+++ tests/Zeus.Server.Tests/PsHwPeakResolutionTests.cs
@@ -19,6 +19,7 @@
 // (https://github.com/dl1bz/deskhpsdr), maintained by Heiko (DL1BZ).
 // Both are GPL-2.0-or-later.
 
+using Zeus.Contracts;
 using Zeus.Protocol1.Discovery;
 using Zeus.Server;
 using Xunit;
